+++ Cavokator/Resources/layout/wx_weather_main.axml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/weather_fragment_container"
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
     android:fillViewport="true"
